+++ cfg_data.lua
@@ -1,14 +1,14 @@
 local lib_sources = {'base64.c', 'core.c', 'coretests.c', 'cpuinfo.c',
     'blake2s.c', 'entropy.c',
-    'extratests.c', 'fileio.c', 'lineardep.c', 'hwtests.c', 'specfuncs.c',
-    'threads_intf.c'}
+    'extratests.c', 'fileio.c', 'lfsr_period.c', 'lineardep.c', 'hwtests.c',
+    'specfuncs.c', 'threads_intf.c'}
 
 local bat_sources = {'bat_express.c', 'bat_brief.c', 'bat_default.c',
     'bat_file.c', 'bat_full.c', 'bat_special.c'}
 
 local lib_headers = {'apidefs.h', 'cinterface.h', 'base64.h', 'blake2s.h', 'core.h',
     'coredefs.h', 'coretests.h', 'cpuinfo.h', 'entropy.h', 'extratests.h', 'fileio.h',
-    'hwtests.h', 'int128defs.h', 'lineardep.h', 'specfuncs.h',
+    'hwtests.h', 'int128defs.h', 'lfsr_period.h', 'lineardep.h', 'specfuncs.h',
     'threads_intf.h', 'version.h', 'x86exts.h'}