+++ include/smokerand/lfsr_period.h
@@ -14,5 +14,20 @@
 #include "smokerand/core.h"
 #endif
 
+typedef struct {
+    int check_validity;
+} LfsrPeriodOptions;
+
+
+typedef enum {
+    LFSR_PERIOD_MAX     = 0,
+    LFSR_PERIOD_NOT_MAX = 1,
+    LFSR_PERIOD_ERROR   = 2
+} LfsrPeriodResult;
+
+
+LfsrPeriodResult lfsr_period_test(const GeneratorInfo *gen, const CallerAPI *intf,
+    const LfsrPeriodOptions *opts);
+
 BatteryExitCode battery_lfsr_period(const GeneratorInfo *gen, const CallerAPI *intf,
     const BatteryOptions *opts);